post-release fixes: 32/64bit problem and GDL requirements
yet another instance of that obnoxious problem that "long"
is just 32bit on i386 platforms. Why the hell does such
a broken type get the preference of convenient notation??

found while building packages on Debian/Jessie 32bit

moreover, we lower the GDL requirement level, which was
chose more or less arbitrary anyway, to support also
building on Mint (and Ubuntu Trusty). Hopefully this
does not lead to further problems elsewhere